+++ CHANGELOG.md
@@ -1,12 +1,13 @@
-DevOps Tool: WordPress Platform Support Changelog
+Conductor: WordPress Platform Support Changelog
 ==============================================
 
 # 0.2.0 (Unreleased)
+- Renamed to Conductor
 - Updated PHP version requirement to 7.1
-- Updated config to work with DevopsToolAppOrchestration
+- Updated config to work with ConductorAppOrchestration
 
 # 0.1.1
 - Added zf config-provider Composer setting
 
 # 0.1.0
-- Initial build copied over from DevOps tool
+- Initial build copied over from Conductor